cpu/nrf5x: provide gpio_t type definition

+++ boards/particle-argon/include/periph_conf.h
@@ -36,8 +36,8 @@ static const uart_conf_t uart_config[] = {
         .dev        = NRF_UARTE0,
         .rx_pin     = GPIO_PIN(0,8),
         .tx_pin     = GPIO_PIN(0,6),
-        .rts_pin    = (uint8_t)GPIO_UNDEF,
-        .cts_pin    = (uint8_t)GPIO_UNDEF,
+        .rts_pin    = GPIO_UNDEF,
+        .cts_pin    = GPIO_UNDEF,
         .irqn       = UARTE0_UART0_IRQn,
     },
     {
